+++ spring-boot-project/spring-boot-actuator-autoconfigure/src/main/java/org/springframework/boot/actuate/autoconfigure/metrics/export/properties/StepRegistryPropertiesConfigAdapter.java
@@ -16,8 +16,6 @@
 
 package org.springframework.boot.actuate.autoconfigure.metrics.export.properties;
 
-import java.time.Duration;
-
 import io.micrometer.core.instrument.step.StepRegistryConfig;
 
 /**
@@ -30,40 +28,10 @@ import io.micrometer.core.instrument.step.StepRegistryConfig;
  * @since 2.0.0
  */
 public abstract class StepRegistryPropertiesConfigAdapter<T extends StepRegistryProperties>
-		extends PropertiesConfigAdapter<T> implements StepRegistryConfig {
+		extends PushRegistryPropertiesConfigAdapter<T> {
 
 	public StepRegistryPropertiesConfigAdapter(T properties) {
 		super(properties);
-	}
-
-	@Override
-	public String prefix() {
-		return null;
-	}
-
-	@Override
-	public String get(String k) {
-		return null;
-	}
-
-	@Override
-	public Duration step() {
-		return get(T::getStep, StepRegistryConfig.super::step);
-	}
-
-	@Override
-	public boolean enabled() {
-		return get(T::isEnabled, StepRegistryConfig.super::enabled);
-	}
-
-	@Override
-	public int numThreads() {
-		return get(T::getNumThreads, StepRegistryConfig.super::numThreads);
-	}
-
-	@Override
-	public int batchSize() {
-		return get(T::getBatchSize, StepRegistryConfig.super::batchSize);
 	}
 
 }
