+++ spring-boot-project/spring-boot-actuator-autoconfigure/src/test/java/org/springframework/boot/actuate/autoconfigure/metrics/export/properties/StepRegistryPropertiesTests.java
@@ -18,23 +18,15 @@ package org.springframework.boot.actuate.autoconfigure.metrics.export.properties
 
 import io.micrometer.core.instrument.step.StepRegistryConfig;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
 /**
  * Base tests for {@link StepRegistryProperties} implementation.
  *
  * @author Stephane Nicoll
  */
-public abstract class StepRegistryPropertiesTests {
+public abstract class StepRegistryPropertiesTests extends PushRegistryPropertiesTests {
 
-	@SuppressWarnings("deprecation")
 	protected void assertStepRegistryDefaultValues(StepRegistryProperties properties, StepRegistryConfig config) {
-		assertThat(properties.getStep()).isEqualTo(config.step());
-		assertThat(properties.isEnabled()).isEqualTo(config.enabled());
-		assertThat(properties.getConnectTimeout()).isEqualTo(config.connectTimeout());
-		assertThat(properties.getReadTimeout()).isEqualTo(config.readTimeout());
-		assertThat(properties.getNumThreads()).isEqualTo(config.numThreads());
-		assertThat(properties.getBatchSize()).isEqualTo(config.batchSize());
+		super.assertStepRegistryDefaultValues(properties, config);
 	}
 
 }
